new: usr: Dockerized application
