+++ app/thlib/__init__.py
@@ -1,0 +1,2 @@
+from bcolors import BColors  # NOQA
+from utility import Utility  # NOQA
